+++ .gitignore
@@ -3,3 +3,4 @@
 /docker/mysql/volumes/
 /log/
 /.env
+/.vscode
